Recuperation de la string pour le return - avancement majeure - Makefile - Filechecker 88/197 28/40 48

+++ ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: clegoube <clegoube@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 12:53:46 by clegoube          #+#    #+#             */
-/*   Updated: 2017/01/16 16:28:08 by clegoube         ###   ########.fr       */
+/*   Updated: 2017/01/30 18:52:02 by clegoube         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-static size_t	ft_length_number(intmax_t nb)
+static int	ft_length_number(intmax_t nb)
 {
 	int			i;
 
@@ -32,7 +32,7 @@ static size_t	ft_length_number(intmax_t nb)
 char			*ft_itoa(intmax_t n)
 {
 	char			*string;
-	unsigned int	nb;
+	uintmax_t		nb;
 	size_t			i;
 
 	i = ft_length_number(n);
@@ -55,5 +55,6 @@ char			*ft_itoa(intmax_t n)
 		string[i--] = (nb % 10) + '0';
 		nb = nb / 10;
 	}
+
 	return (string);
 }
